third_party/WebKit/LayoutTests/http/tests/serviceworker/serviceworkerglobalscope-scope.html

<!DOCTYPE html>
<title>ServiceWorkerGlobalScope: scope property</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
function scope_test(name, scope) {
  var t = async_test('Verify the scope property: ' + name);
  t.step(function() {
      var expectedScope, options;
      if (scope) {
        expectedScope = new URL(scope, document.location).toString();
      } else {
        expectedScope = new URL('/', document.location).toString();
      }

      var registration;
      var script = 'resources/serviceworkerglobalscope-scope-worker.js';
      var options = scope ? { scope: scope } : {};
      navigator.serviceWorker.unregister(scope)
        .then(function() {
            return navigator.serviceWorker.register(script, options);
          })
        .then(function(r) {
            registration = r;
            return wait_for_update(t, registration);
          })
        .then(function(worker) {
            return new Promise(function(resolve) {
                var messageChannel = new MessageChannel();
                messageChannel.port1.onmessage = resolve;
                worker.postMessage('', [messageChannel.port2]);
              });
          })
        .then(function(e) {
            var message = e.data;
            assert_equals(message.initialScope, expectedScope,
                          'Worker should see the scope on eval.');
            assert_equals(message.currentScope, expectedScope,
                          'Worker scope should not change.');
            return registration.unregister();
          })
        .then(function() {
            t.done();
          })
        .catch(unreached_rejection(t));
    });
}

scope_test('default');
scope_test('relative path', '/a/b/c/');
scope_test('absolute url', 'http://127.0.0.1:8000/');

</script>
